requirements.txt
python-dotenv>=0.19.0
pandas==2.2.3
yfinance==0.2.54
alpaca-py==0.42.2
apscheduler==3.11.0
aiogram==3.18.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0